Updated nibble to byte
From my understanding, a nibble is only 4-bits and represented by 1 hexadecimal character, while a byte is 8-bits and represented by 2 hexadecimal characters. Based on my understanding of this code, it takes a 2-byte number and operates on the least significant and most significant bytes of the 2-byte number from the user. Because of this, nibble should probably be replaced by byte to convey the correct size of the values being operated on.